+++ frontend/context/XContext.js
@@ -1,6 +1,7 @@
 import { createContext, useEffect, useState } from 'react'
 import { useRouter } from 'next/router'
 import { abi } from '../constants'
+import { ethers } from 'ethers'
 
 export const XContext = createContext()
 
@@ -8,10 +9,11 @@ export const XProvider = ({ children }) => {
   const [appStatus, setAppStatus] = useState('')
   const [currentAccount, setCurrentAccount] = useState('')
   const [currentUser, setCurrentUser] = useState({})
+  const [me, setMe] = useState()
 
   const router = useRouter()
 
-  const xtelptAddress = "0x2e60513162fa4c9a324396b98ed0141e66138da0"
+  const xtelptAddress = "0x7aD928627634055f32129C1bf22AB689006EF293"
 
 
   useEffect(() => {
@@ -59,6 +61,38 @@ export const XProvider = ({ children }) => {
       setAppStatus('error')
     }
   }
+
+
+  const updateUIValues = async () => {
+    const provider = new ethers.providers.Web3Provider(window.ethereum)
+    const xtelptContract = new ethers.Contract(xtelptAddress, abi, provider)
+
+    try {
+      const addressArray = await window.ethereum.request({
+        method: 'eth_accounts',
+      })
+      if (addressArray.length > 0) {
+        let prof = await xtelptContract.getProfile(`${addressArray[0]}`)
+        setMe(prof)
+        console.log(me)
+      } else {
+        console.log("no metamask")
+      }
+    } catch (err) {
+      console.log("an error occured")
+    }
+
+  }
+
+  useEffect(() => {
+    setTimeout(() => {
+      updateUIValues()
+    }, 300);
+  }, [])
+
+
+
+
   return (
     <XContext.Provider
       value={{
@@ -67,7 +101,8 @@ export const XProvider = ({ children }) => {
         connectWallet,
         setAppStatus,
         xtelptAddress,
-        abi
+        abi,
+        me
       }}
     >
       {children}
